+++ eems/scripts.py
@@ -62,7 +62,7 @@ def main():
         c.prepare()
 
     elif args.command == 'read':
-        t = ds18b20.Temp()
+        t = ds18b20.Temp(console=True)
         t.read()
 
     elif args.command == 'monitor':
@@ -85,7 +85,7 @@ def main():
         if args.interval:
             interval = args.interval
         else:
-            interval = None
+            interval = 60
         if args.duration:
             duration = args.duration
         else: